+++ src/engine/client/srvbrowse.cpp
@@ -435,9 +435,9 @@ void CServerBrowser::Set(const NETADDR &Addr, int Type, int Token, const CServer
 		{
 			SetInfo(pEntry, *pInfo);
 			if(m_ServerlistType == IServerBrowser::TYPE_LAN)
-				pEntry->m_Info.m_Latency = (time_get()-m_BroadcastTime)*1000/time_freq();
+				pEntry->m_Info.m_Latency = (int)((time_get()-m_BroadcastTime)*1000/time_freq());
 			else
-				pEntry->m_Info.m_Latency = (time_get()-pEntry->m_RequestTime)*1000/time_freq();
+				pEntry->m_Info.m_Latency = (int)((time_get()-pEntry->m_RequestTime)*1000/time_freq());
 			RemoveRequest(pEntry);
 		}
 	}
@@ -449,9 +449,9 @@ void CServerBrowser::Set(const NETADDR &Addr, int Type, int Token, const CServer
 			SetInfo(pEntry, *pInfo);
 
 			if(m_ServerlistType == IServerBrowser::TYPE_LAN)
-				pEntry->m_Info.m_Latency = (time_get()-m_BroadcastTime)*1000/time_freq();
+				pEntry->m_Info.m_Latency = (int)((time_get()-m_BroadcastTime)*1000/time_freq());
 			else
-				pEntry->m_Info.m_Latency = (time_get()-pEntry->m_RequestTime)*1000/time_freq();
+				pEntry->m_Info.m_Latency = (int)((time_get()-pEntry->m_RequestTime)*1000/time_freq());
 			RemoveRequest(pEntry);
 		}
 	}
