+++ src/game/client/components/effects.cpp
@@ -185,8 +185,8 @@ void CEffects::Explosion(vec2 Pos)
 			if(x == 0 && y == 0)
 				continue;
 			
-			float a = 1 - (length(vec2(x,y)) / length(vec2(8,8)));
-			m_pClient->m_pFlow->Add(Pos+vec2(x,y)*16, normalize(vec2(x,y))*5000.0f*a, 10.0f);
+			float a = 1 - (length(vec2((float)x,(float)y)) / length(vec2(8,8)));
+			m_pClient->m_pFlow->Add(Pos+vec2((float)x,(float)y)*16, normalize(vec2((float)x,(float)y))*5000.0f*a, 10.0f);
 		}
 		
 	// add the explosion
